+++ docs/devref/mdpo/mdpo.compat.rst
@@ -1,0 +1,6 @@
+*********
+compat.py
+*********
+
+.. automodule:: mdpo.compat
+   :members:
