--- docs/devref/mdpo/mdpo.context.rst
@@ -1,6 +1,0 @@
-**********
-context.py
-**********
-
-.. automodule:: mdpo.context
-   :members:
