+++ mdpo/mdpo2html/__main__.py
@@ -17,7 +17,7 @@ from mdpo.cli import (
     cli_codespan,
     parse_command_aliases_cli_arguments,
 )
-from mdpo.context import environ
+from mdpo.io import environ
 from mdpo.mdpo2html import MdPo2HTML
 
 
@@ -113,7 +113,7 @@ def run(args=[]):
 
 
 def main():
-    sys.exit(run(args=sys.argv[1:])[1])  # pragma: no cover
+    raise SystemExit(run(args=sys.argv[1:])[1])  # pragma: no cover
 
 
 if __name__ == '__main__':
